check likes and dislikes middleware created
